+++ tools/validate.js
@@ -2,7 +2,7 @@
  * Validate whether all functions in math.js are documented in math.expression.docs
  */
 var gutil = require('gulp-util'),
-    math = require('../index')(),
+    math = require('../index'),
     prop;
 
 // names to ignore