+++ HISTORY.md
@@ -7,6 +7,8 @@
 - Implemented support for `null` in all functions.
 - Implemented support for "raw" functions in the expression parser. Raw 
   functions are invoked with unevaluated parameters (nodes).
+- Expressions in the expression parser can now be spread over multiple lines,
+  like '2 +\n3'.
 - Changed default value of the option `wrap` of function `math.import` to false.
 - Changed the default value for new entries in a resized matrix when to zero. 
   To leave new entries uninitialized, use the new constant `math.uninitialized`